+++ src/Microsoft.Health.Fhir.CosmosDb/Features/Search/Queries/QueryBuilder.cs
@@ -5,6 +5,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 using EnsureThat;
 using Microsoft.Azure.Cosmos;
@@ -81,7 +82,7 @@ namespace Microsoft.Health.Fhir.CosmosDb.Features.Search.Queries
                     searchOptions.Expression.AcceptVisitor(expressionQueryBuilder);
                 }
 
-                AppendLiteralFilterCondition(
+                AppendFilterCondition(
                    "AND",
                    true,
                    (KnownResourceWrapperProperties.IsHistory, false),
@@ -252,11 +253,6 @@ namespace Microsoft.Health.Fhir.CosmosDb.Features.Search.Queries
             private void AppendFilterCondition(string logicalOperator, bool equal, params (string, object)[] conditions)
             {
                 _queryHelper.AppendFilterCondition(logicalOperator, equal, conditions);
-            }
-
-            private void AppendLiteralFilterCondition(string logicalOperator, bool equal, params (string, object)[] conditions)
-            {
-                _queryHelper.AppendLiteralFilterCondition(logicalOperator, equal, conditions);
             }
 
             private void AppendFilterCondition(string name, object value, bool equal)
